Move confirmation mode feature from LLM settings to Application settings
- Moved confirmation mode toggle from LLM settings page to Application settings page
- Moved security analyzer dropdown to Application settings (shown when confirmation mode enabled)
- Maintained identical functionality and behavior
- Cleaned up llm-settings.tsx by removing 186 lines of related code
- Added 173 lines to app-settings.tsx with proper state management and UI components
